Moved static buildIcon method to playerCreateView
